Update help pages and add features
This commit updates the help pages for Goal Planner and Medical Records & Reminders.  It also adds a camera option to the diary for taking pictures and implements voice-to-text functionality for all note/text fields.  Finally, it updates the My Calendar's Events and Add New Event dropdowns to include "Goals" and "Medical".